requirements.txt
jax
orbax
seqio
einops
absl-py
clu
t5
tensorboard
# Following packages are only required for running simply/utils/math_eval.py
sympy
pylatexenc
pytest